Use bf16-mixed during training

+++ train_psi_predictor.py
@@ -181,7 +181,7 @@ def main():
             # checkpointing_cb,
             # early_stopping_cb,
         ],
-        precision="32-true",
+        precision="bf16-mixed",
         strategy="ddp",
         reload_dataloaders_every_n_epochs=1,
     )
